20190823 18:00 근목

+++ final_project/src/main/java/com/kh/spring/entity/ReviewDto.java
@@ -14,7 +14,7 @@ public class ReviewDto {
 	private String title; // 제목
 	private String content; // 내용
 	private int read; //조회수
-	private int score; // 평점
+	private double score; // 평점
 	private String write_status; // 작성상태
 	private String reply_status; // 답변상태
 	private String regist_date; //등록일
